+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 2.2.1-SNAPSHOT
 
+* [ENHANCEMENT] Improved error handling with Prometheus API
+* [BUG] Fixed (#18)
+
 ## 2.2.0 / 2022-01-29
 
 * [FEATURE] Support for Prometheus `query_range` API.